Expose UserService as external API
